+++ scanpointgenerator/generators/__init__.py
@@ -18,3 +18,4 @@ from scanpointgenerator.generators.linegenerator import LineGenerator
 from scanpointgenerator.generators.lissajousgenerator import LissajousGenerator
 from scanpointgenerator.generators.spiralgenerator import SpiralGenerator
 from scanpointgenerator.generators.staticpointgenerator import StaticPointGenerator
+from scanpointgenerator.generators.zipgenerator import ZipGenerator